backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
mysql-connector-python==8.0.33
pandas==2.1.1
matplotlib==3.7.3
numpy==1.26.0